requirements.txt
streamlit==1.32.2
tensorflow==2.15.0
numpy==1.26.4
pillow